Refactor widgets
