+++ androidHyperskillApp/src/main/java/org/hyperskill/app/android/challenge/ui/ChallengeCardDefaults.kt
@@ -1,0 +1,8 @@
+package org.hyperskill.app.android.challenge.ui
+
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.ui.unit.dp
+
+object ChallengeCardDefaults {
+    val paddingValues: PaddingValues = PaddingValues(20.dp)
+}
