+++ androidHyperskillApp/src/main/java/org/hyperskill/app/android/challenge/ui/ChallengeScaffold.kt
@@ -1,0 +1,26 @@
+package org.hyperskill.app.android.challenge.ui
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.ColumnScope
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import org.hyperskill.app.android.core.view.ui.widget.compose.HyperskillCard
+
+@Composable
+fun ChallengeScaffold(
+    modifier: Modifier = Modifier,
+    content: @Composable ColumnScope.() -> Unit
+) {
+    HyperskillCard(
+        contentPadding = ChallengeCardDefaults.paddingValues,
+        onClick = {},
+        modifier = modifier
+    ) {
+        Column(
+            content = content,
+            verticalArrangement = Arrangement.spacedBy(16.dp)
+        )
+    }
+}
